requirements.txt
numpy>=1.21.2
matplotlib>=3.4.3
scipy>=1.7.1
numpy-stl>=2.16.3
PyQt5>=5.15.6
pyqtgraph>=0.12.4
pyopengl>=3.1.6